Clean up AbstractKeyValueService, put complexity into DbKvs
There's a method (on a hot path) in AbstractKeyValueService called
multiPut. It isn't used on any codepath except in DbKvs, and
provides no benefit except in there.

Remove it from AbstractKeyValueService, push into DbKvs. Reimplement
more efficiently in a way that avoids tests all creating 16 threads
per test.
