+++ atlasdb-client/src/main/java/com/palantir/atlasdb/keyvalue/impl/InMemoryKeyValueService.java
@@ -48,6 +48,7 @@ import com.google.common.collect.Sets;
 import com.google.common.io.BaseEncoding;
 import com.google.common.primitives.Longs;
 import com.google.common.primitives.UnsignedBytes;
+import com.google.common.util.concurrent.MoreExecutors;
 import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.keyvalue.api.BatchColumnRangeSelection;
 import com.palantir.atlasdb.keyvalue.api.CandidateCellForSweeping;
@@ -83,8 +84,7 @@ public class InMemoryKeyValueService extends AbstractKeyValueService {
     private final boolean createTablesAutomatically;
 
     public InMemoryKeyValueService(boolean createTablesAutomatically) {
-        this(createTablesAutomatically,
-                PTExecutors.newFixedThreadPool(16, PTExecutors.newNamedThreadFactory(true)));
+        this(createTablesAutomatically, MoreExecutors.newDirectExecutorService());
     }
 
     public InMemoryKeyValueService(boolean createTablesAutomatically,
@@ -376,6 +376,11 @@ public class InMemoryKeyValueService extends AbstractKeyValueService {
     }
 
     @Override
+    public void multiPut(Map<TableReference, ? extends Map<Cell, byte[]>> valuesByTable, final long timestamp) {
+        valuesByTable.forEach((tableRef, values) -> put(tableRef, values, timestamp));
+    }
+
+    @Override
     public void put(TableReference tableRef, Map<Cell, byte[]> values, long timestamp) {
         putInternal(tableRef, KeyValueServices.toConstantTimestampValues(values.entrySet(), timestamp), false);
     }
@@ -481,6 +486,11 @@ public class InMemoryKeyValueService extends AbstractKeyValueService {
     public void dropTable(TableReference tableRef) {
         tables.remove(tableRef);
         tableMetadata.remove(tableRef);
+    }
+
+    @Override
+    public void truncateTables(final Set<TableReference> tableRefs) {
+        tableRefs.forEach(this::truncateTable);
     }
 
     @Override
