+++ docs/source/release_notes/release-notes.rst
@@ -54,6 +54,10 @@ develop
          - If we make a successful request to a Cassandra client, we now remove it from the overall Cassandra service's blacklist.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/3156>`__)
 
+    *    - |improved|
+         - Move a complicated and elsewhere overridden method from AbstractKeyValueService into DbKvsKeyValueService
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/3159>`__)
+
     *    - |fixed|
          - The (Thrift-backed) ``CassandraKeyValueService`` now returns correctly for CQL queries that return null.
            Previously, they would throw an exception when we attempted to log information about the response.
